requirements.txt
python-dotenv
schedule
Pillow
mss
firebase-admin
google-generativeai